+++ src/spring/HISTORY.md
@@ -3,6 +3,7 @@ Release History
 
 1.1.8
 ---
+* Add command `az spring app connect`.
 * Add the parameter `language_framework` for deploying the customer image app.
 
 1.1.7